backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for the Histora platform"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "psycopg[binary]>=3.3.3",
]
